requirements.txt
# pip install -r requirements.txt
blueprint==3.4.2
Flask-JWT-Extended==3.14.0
Flask-Cors==3.0.9
google-cloud-core==1.4.1
google-cloud-datastore==1.15.0
google-cloud-ndb==1.5.1
PyJWT==1.6.4
schedule==1.1.0
skpy==0.10.4
flask
flask-restx
grpcio
Werkzeug